requirements.txt
numpy
pandas
scikit-learn
python-dotenv
pyserial
joblib